Move the core logic to kotlin

+++ src/main/java/cl/bozz/nqueensdfs/utils/TotalPermutationsCalculator.kt
@@ -1,9 +1,9 @@
 package cl.bozz.nqueensdfs.utils
 
 object TotalPermutationsCalculator {
-    fun totalPermutations(boardSize: Long): Long {
+    fun totalPermutations(boardSize: Int): Long {
         val totalCells = boardSize * boardSize
-        var result = totalCells
+        var result = totalCells.toLong()
         for (i in 1 until boardSize) {
             val mult = totalCells - i
             result *= mult
